requirements.txt
aiohttp==3.11.7
bleak==0.22.3
PyYAML==6.0.2
